modified:   Cfg/Configuration.h 	modified:   Entity/Movies.cpp 	modified:   Entity/Movies.h 	modified:   Makefile 	modified:   Model/SparseMatri/SparseMatriOpt.cpp 	modified:   Model/SparseMatri/SparseMatriOpt.h 	modified:   Pretreatment.cpp 	modified:   Pretreatment.h 	new file:   Service/MoviesReader.cpp 	new file:   Service/MoviesReader.h 	modified:   Service/RatingReader.cpp 	modified:   test.cpp

+++ Model/SparseMatri/SparseMatriOpt.cpp
@@ -12,7 +12,7 @@ int SpMatrix::get(int x, int y)
 {
 	return SpMatrix::mat.coeffRef(x,y);
 }
-void SpMatrix::set(int x, int y, int val)
+void SpMatrix::set(int x, int y, short val)
 {
 	SpMatrix::mat.insert(x,y) = val;
 }
